requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
twilio==8.10.0
openai==1.3.7
pytesseract==0.3.10
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
celery==5.3.4
redis==5.0.1